refactor: rename SparseVec and TypeErasedSparseVec
