+++ src/collections.rs
@@ -1,7 +1,7 @@
 mod generational_vec;
-mod type_erased_array;
-mod typed_array;
+mod sparse_vec;
+mod type_erased_sparse_vec;
 
 pub use generational_vec::*;
-pub use type_erased_array::*;
-pub use typed_array::*;
+pub use sparse_vec::*;
+pub use type_erased_sparse_vec::*;
